requirements.txt
requests==2.22.0
BeautifulSoup4==4.7.1
python-telegram-bot==11.1.0